Move config functions and variables to separate module
- Add config module.
- Use `add_section` rather than declaring an empty `Options` section to
  avoid potentially overwriting one.
- Resolves #36.
